examples/raspberry_pi_zero_2w/requirements.txt
# Enhanced LoRa SX126x Sensor Node Requirements
# Version 2.0.0

# Core dependencies
cryptography>=3.4.8
psutil>=5.8.0
jsonschema>=4.0.0

# Hardware interfaces
RPi.GPIO>=0.7.1
spidev>=3.5
smbus2>=0.4.1

# Sensor libraries
adafruit-circuitpython-dht>=3.7.0
adafruit-circuitpython-bmp280>=3.2.0
w1thermsensor>=2.0.0

# Data handling
numpy>=1.21.0
pandas>=1.3.0

# Networking and communication
pyserial>=3.5
paho-mqtt>=1.6.0

# Utilities
click>=8.0.0
colorama>=0.4.4
tqdm>=4.62.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.18.0
black>=21.0.0
flake8>=4.0.0

# Optional: Advanced features
# scipy>=1.7.0  # For advanced signal processing
# matplotlib>=3.4.0  # For data visualization
# influxdb-client>=1.24.0  # For InfluxDB integration
